[Silver II] Title: 랜선 자르기, Time: 4 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/1654. 랜선 자르기/랜선 자르기.cc
@@ -7,7 +7,7 @@ int main() {
 	cin.tie(NULL);
 	cout.tie(NULL);
 	int k, n, lan[10000];
-	long long max = 0;
+	int max = 0;
 	cin >> k >> n;
 	for (int i = 0; i < k; i++)
 	{
